fix: semantic_similarity returns numpy.float32, breaking eval result storage
sentence-transformers' model.encode() returns numpy arrays, so the cosine
similarity calculation stayed a numpy.float32 all the way to the return
value. eval_service persists EvalResult.scores as JSON, and json.dumps()
can't serialize numpy scalars - every row of every eval run using this
metric failed with "Object of type float32 is not JSON serializable".

Caught by actually running an eval end-to-end against a live Ollama model
in the docker-compose smoke test; the unit tests never exercised this path
because sentence-transformers wasn't installed in the dev venv despite
being in requirements.txt, so _get_model() was never really called.

+++ backend/evaluators/semantic_similarity.py
@@ -32,4 +32,6 @@ class SemanticSimilarityEvaluator(Evaluator):
         model = _get_model()
         emb_output, emb_expected = model.encode([output, expected_output])
         similarity = _cosine(emb_output, emb_expected)
-        return max(0.0, min(1.0, (similarity + 1) / 2))
+        # model.encode() returns numpy arrays, so _cosine's result is numpy.float32 -
+        # not JSON-serializable when eval_service persists scores to the DB.
+        return float(max(0.0, min(1.0, (similarity + 1) / 2)))
